Move all variations of NameCalebrations to the same package
