+++ memento/src/main/java/com/alexstyl/specialdates/events/namedays/MutableNameCelebrations.kt
@@ -1,0 +1,7 @@
+package com.alexstyl.specialdates.events.namedays
+
+import com.alexstyl.specialdates.date.Date
+
+interface MutableNameCelebrations : NameCelebrations {
+    fun addDate(date: Date)
+}
