+++ memento/src/main/java/com/alexstyl/specialdates/events/namedays/calendar/resource/CharacterNode.kt
@@ -4,6 +4,8 @@ import com.alexstyl.gsc.Index
 import com.alexstyl.specialdates.date.Date
 import com.alexstyl.specialdates.events.namedays.NameCelebrations
 import com.alexstyl.specialdates.events.namedays.NoNameCelebrations
+import com.alexstyl.specialdates.events.namedays.ArrayNameCelebrations
+import com.alexstyl.specialdates.events.namedays.MutableNameCelebrations
 import java.text.Collator
 
 class CharacterNode(private val keySound: Char?) : Node {
